requirements.txt
asgiref==3.8.1
boto3==1.34.145
botocore==1.34.145
Django==5.0.7
django-filter==24.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.1
jmespath==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
